+++ packages/auth-provider-email/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.0](https://github.com/activescott/auth/compare/auth-provider-email@1.0.3...auth-provider-email@1.1.0) (2026-08-07)
+
+### Features
+
+* ship CaptureEmailTransport for e2e tests ([451169b](https://github.com/activescott/auth/commit/451169b8d49fe550378600e123360c8a192ddd2d))
+
+## [1.0.3](https://github.com/activescott/auth/compare/auth-provider-email@1.0.2...auth-provider-email@1.0.3) (2026-08-06)
+
+### Bug Fixes
+
+* throttle sends per recipient address ([df1dcad](https://github.com/activescott/auth/commit/df1dcadef135425571b8fe914b32ac4279b1a966)), closes [#56](https://github.com/activescott/auth/issues/56)
+
 ## [1.0.2](https://github.com/activescott/auth/compare/auth-provider-email@1.0.1...auth-provider-email@1.0.2) (2026-08-02)
 
 ### Bug Fixes